refactor(audio-analyzer, web-request-interceptor): enhance audio message detection and duration extraction
- 在 audio-analyzer.js 中新增 isVoiceMessage 函數，強化對語音訊息的檢測邏輯，並整合對請求元數據的檢查。
- 更新 web-request-interceptor.js，將對音訊持續時間的獲取邏輯整合至新的 audio-analyzer 模組，提升代碼的可讀性與維護性。
- 調整日誌記錄，提供更詳細的音訊請求處理資訊，增強調試能力。
- 此變更有助於提升音訊請求的準確性與穩定性，並為未來的功能擴展提供更清晰的基礎。
